moved spawner to the sprites folder
